+++ app/javascript/components/Common/Table/ColumnFilter/index.js
@@ -1,0 +1,1 @@
+export { default as InputFilter } from "./InputFilter";
